refactor(pairs): one shared pairing engine for POS + cart upsell
Extract the 'Pair with a Bite' scoring (combo/roundPromo/co-purchase/usual/
round/complement) into @celsius/shared suggestPairs(), parameterized by
channel + injected supabase client. The POS route and the customer in-cart
upsell are now thin wrappers over it (channel 'pos' vs 'pickup'), so the two
surfaces share weights (app_settings.pair_weights), combo/promo awareness,
and the nightly tuner — no more drift between two engines.
